+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - [IMPROVEMENT] Now using logging for debug output.
 - [IMPROVEMENT] Simplified marshalling.
 - [IMPROVEMENT] Add CSEBase to Session Resource. 
+- [IMPROVEMENT] Add Originator to Resources.
+
 
 ## Version 0.7 (2018-05-13)
 - [IMPROVEMENT] Added support for &lt;remoteCSE> resource type.